+++ src/commonMain/kotlin/io/klogging/config/Defaults.kt
@@ -18,8 +18,8 @@
 
 package io.klogging.config
 
+import io.klogging.Level
 import io.klogging.dispatching.STDOUT
-import io.klogging.events.Level
 import io.klogging.rendering.RENDER_SIMPLE
 
 /** Simple sink configuration for rendering simple strings to the standard output stream. */
