+++ src/commonMain/kotlin/io/klogging/impl/NoCoLoggerImpl.kt
@@ -18,9 +18,9 @@
 
 package io.klogging.impl
 
+import io.klogging.Level
 import io.klogging.Logging
 import io.klogging.NoCoLogger
-import io.klogging.events.Level
 import io.klogging.events.LogEvent
 import io.klogging.template.templateItems
 import kotlinx.coroutines.CoroutineScope
