+++ src/jvmTest/kotlin/io/klogging/KloggerTest.kt
@@ -18,7 +18,6 @@
 
 package io.klogging
 
-import io.klogging.events.Level
 import io.klogging.events.LogEvent
 import io.klogging.events.hostname
 import io.klogging.template.templateItems
