+++ src/jvmTest/kotlin/io/klogging/Playpen.kt
@@ -22,7 +22,6 @@ import io.klogging.config.loggingConfiguration
 import io.klogging.config.seq
 import io.klogging.context.logContext
 import io.klogging.dispatching.STDOUT
-import io.klogging.events.Level
 import io.klogging.rendering.RENDER_SIMPLE
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
