+++ src/jvmTest/kotlin/io/klogging/config/JsonConfigurationTest.kt
@@ -18,8 +18,8 @@
 
 package io.klogging.config
 
+import io.klogging.Level
 import io.klogging.dispatching.STDOUT
-import io.klogging.events.Level
 import io.klogging.rendering.RENDER_SIMPLE
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.collections.shouldContainExactly
