+++ src/jvmTest/kotlin/io/klogging/rendering/RenderGelfTest.kt
@@ -18,7 +18,7 @@
 
 package io.klogging.rendering
 
-import io.klogging.events.Level
+import io.klogging.Level
 import io.klogging.events.LogEvent
 import io.klogging.randomString
 import io.klogging.timestampNow
